refactor: split src/memory.rs into modules
